Implemented previous printMatrix function from rasmus.c with minor tweaks.

+++ held-karp.h
@@ -7,7 +7,7 @@
 
 #include "a-star.h"
 
-int held_karp(double main_node_distances[NUM_MAIN_NODES][NUM_MAIN_NODES], int mask, const int pos, unsigned int iteration);
+double held_karp(double main_node_distances[NUM_MAIN_NODES][NUM_MAIN_NODES], int mask, const int pos, unsigned int iteration);
 void print_held_karp_route(int pos, char between[]);
 
 #endif //HELD_KARP_H
